test float

+++ test/test_pose_graph.cpp
@@ -14,18 +14,18 @@ using namespace mmpilot;
 
 int main()
 {
-	PoseGraphXY<double> graph;
+	PoseGraphXY<float> graph;
 
 	const int N = 6;
 
 	for(int i = 0; i < N; ++i)
 	{
-		graph.add_node(Vec2d(i, 0));
+		graph.add_node(Vec2f(i, 0));
 		if(i > 0) {
-			graph.add_edge(i - 1, i, Vec2d(1, 0));
+			graph.add_edge(i - 1, i, Vec2f(1, 0));
 		}
 	}
-	graph.add_edge(0, N - 1, Vec2d(N, 1), 2);
+	graph.add_edge(0, N - 1, Vec2f(N, 1), 2);
 
 	for(const auto& node : graph.nodes()) {
 		std::cout << node.transpose() << std::endl;
